Create Endpoint and Root nodes
